more cleanup

+++ src/thing_path.cpp
@@ -208,7 +208,7 @@ void thing_path_shorten(Gamep g, Levelsp v, Levelp l, Thingp t, std::vector< spo
     }
   }
 
-  if (! path.empty() != 0u) {
+  if (static_cast<unsigned int>(! path.empty()) != 0U) {
     if (path[ 0 ] == at) {
       path.erase(path.begin());
     }
